+++ fatiando/gravmag/harvester.py
@@ -29,6 +29,8 @@ data.
 * :func:`~fatiando.gravmag.harvester.loadseeds`: Loads from a JSON file a set
   of (x, y, z) points and physical properties that specify the seeds. Pass
   output to :func:`~fatiando.gravmag.harvester.sow`
+* :func:`~fatiando.gravmag.harvester.weights`: Computes data weights based on
+  the distance to the seeds
 
 **Data types**
 
@@ -66,8 +68,7 @@ doi:10.1190/segam2012-0383.1 [`pdf
 """
 import json
 import bisect
-from math import sqrt, exp
-from itertools import izip
+from math import sqrt
 
 import numpy
 
@@ -567,6 +568,38 @@ class Neighbor(object):
         self.distance = distance
         self.effect = effect
 
+def weights(x, y, seeds, influences, decay=2):
+    """
+    Calculate weights for the data based on the distance to the seeds.
+    Use weights to ignore regions of data outside of the target anomaly.
+
+    Parameters:
+
+    * x, y : 1d arrays
+        The x and y coordinates of the observations
+    * seeds : list
+        List of seeds, as returned by :func:`~fatiando.gravmag.harvester.sow`
+    * influences : list of floats
+        The respective diameter of influence for each seed. Observations
+        outside the influence will have very small weights.
+        A recommended value is aproximately the diameter of the anomaly
+    * decay : float
+        The decay factor for the weights. Low decay factor makes the weights
+        spread out more. High decay factor makes the transition from large
+        weights to low weights more abrupt.
+
+    Returns:
+
+    * weights : 1d array
+        The calculated weights
+
+    """
+    distances = numpy.array([((x - s.x)**2 + (y - s.y)**2)/influence**2
+                            for s, influence in zip(seeds, influences)])
+    # min along axis=0 gets the smallest value from each column
+    weights = numpy.exp(-(distances.min(axis=0)**decay))
+    return weights
+
 class Data(object):
     """
     A container for some potential field data.
@@ -575,7 +608,7 @@ class Data(object):
     to calculate the effect of a single cell.
     """
 
-    def __init__(self, x, y, z, data, influence, seeds, meshtype):
+    def __init__(self, x, y, z, data, weights, meshtype):
         self.x = x
         self.y = y
         self.z = z
@@ -589,22 +622,7 @@ class Data(object):
             self.engine = prism_engine
         if self.meshtype == 'tesseroid':
             self.engine = tesseroid_engine
-        if influence is None:
-            self.weights = 1.
-        else:
-            if seeds is None:
-                raise AttributeError('Need seeds to calculate weights')
-            self.weights = self._calculate_weights(influence, seeds)
-
-    def _calculate_weights(self, influence, seeds):
-        """
-        Calculate data weights based on a radius on influence from the seeds.
-        """
-        tmp = -1./(float(influence)**4)
-        weights = numpy.fromiter((
-            exp(tmp*min((xi - s.x)**2 + (yi - s.y)**2 for s in seeds)**2)
-            for xi, yi in izip(self.x, self.y)), dtype=float)
-        return weights
+        self.weights = weights
 
 class Potential(Data):
     """
@@ -620,22 +638,15 @@ class Potential(Data):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-                 meshtype='prism'):
-        Data.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Data.__init__(self, x, y, z, data, weights, meshtype)
         self.prop = 'density'
         self.effectfunc = self.engine.potential
 
@@ -659,22 +670,15 @@ class Gz(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gz
 
 class Gxx(Potential):
@@ -692,22 +696,15 @@ class Gxx(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gxx
 
 class Gxy(Potential):
@@ -725,22 +722,15 @@ class Gxy(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gxy
 
 class Gxz(Potential):
@@ -758,22 +748,15 @@ class Gxz(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gxz
 
 class Gyy(Potential):
@@ -791,22 +774,15 @@ class Gyy(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gyy
 
 class Gyz(Potential):
@@ -824,22 +800,15 @@ class Gyz(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gyz
 
 class Gzz(Potential):
@@ -857,22 +826,15 @@ class Gzz(Potential):
     * data : 1D array
         The values of the data at the observation points
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, influence=None, seeds=None,
-            meshtype='prism'):
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+    def __init__(self, x, y, z, data, weights=1., meshtype='prism'):
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.gzz
 
 class TotalField(Potential):
@@ -892,26 +854,19 @@ class TotalField(Potential):
     * inc, dec : floats
         The inclination and declination of the inducing field
 
-    * influence : None or float
-        If not None, will weight the data by the horizontal distance to the
-        seeds. *influence* is the radius of influence of the seeds on the data.
-        The larger it is, the farther away form the seeds the data can be and
-        still influence the misfit.
-
-    * seeds : None or list of Seeds
-        The list of seeds used in the inversion (as returned by
-        :func:`~fatiando.gravmag.harvester.sow`).
-        Only needed if using weights (i.e., if *influence* is not None).
+    * weight : float or array
+        The weight of this data set in the misfit function. Pass an array to
+        give weights to each data points or a float to weight the entire misfit
+        function. See function :func:`~fatiando.gravmag.harvester.weights`
 
     """
 
-    def __init__(self, x, y, z, data, inc, dec, influence=None, seeds=None,
-            meshtype='prism'):
+    def __init__(self, x, y, z, data, inc, dec, weights=1., meshtype='prism'):
         if meshtype != 'prism':
             raise AttributeError(
                 "Unsupported mesh type '%s' for total field anomaly."
                 % (meshtype))
-        Potential.__init__(self, x, y, z, data, influence, seeds, meshtype)
+        Potential.__init__(self, x, y, z, data, weights, meshtype)
         self.effectfunc = self.engine.tf
         self.prop = 'magnetization'
         self.inc = inc